Add files via upload
This project works on heart disease detection. Besides blood tests and a chest X-ray, tests to diagnose heart disease can include: Electrocardiogram (ECG or EKG). An ECG is a quick and painless test that records the electrical signals in the heart. It can tell if the heart is beating too fast or too slowly.